+++ CHANGELOG.md
@@ -37,6 +37,23 @@ for further information about branching and tagging conventions.
 - Error message styles on text field
 - Select field arrow position
 
+## [1.7.0] - 2019-06-03
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.7.0
+
+## [1.7.0-rc.4] - 2019-05-31
+#### Fixed
+- Pagination while loading limits on limits management page
+
+## [1.7.0-rc.3] - 2019-05-30
+#### Fixed
+- Fixed Withdrawals screen list and review issues
+
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.7.0-rc.2
+
 ## [1.7.0-rc.2] - 2019-05-21
 ### "Under the hood" changes
 #### Changed
@@ -296,7 +313,10 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.2...HEAD
+[Unreleased]: https://github.com/tokend/admin-panel/compare/1.7.0...HEAD
+[1.7.0]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.4...1.7.0
+[1.7.0-rc.4]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.3...1.7.0-rc.4
+[1.7.0-rc.3]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.2...1.7.0-rc.3
 [1.7.0-rc.2]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.1...1.7.0-rc.2
 [1.7.0-rc.1]: https://github.com/tokend/admin-panel/compare/1.7.0-rc.0...1.7.0-rc.1
 [1.7.0-rc.0]: https://github.com/tokend/admin-panel/compare/1.6.0...1.7.0-rc.0